Change DTO to Request and Response, Test all

+++ API/EasyMall/EasyMall.Services/Implements/CartService.cs
@@ -1,10 +1,9 @@
 ﻿using AutoMapper;
 using EasyMall.DALs.Entities;
 using EasyMall.DALs.Repositories.Interfaces;
-using EasyMall.DTOs.DTOs;
+using EasyMall.Models.DTOs.Request;
+using EasyMall.Models.DTOs.Response;
 using EasyMall.Services.Interfaces;
-using LinqKit;
-using MayNghien.Infrastructure.Request.Base;
 using MayNghien.Models.Response.Base;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
@@ -33,29 +32,25 @@ namespace EasyMall.Services.Implements
             _productPriceRepository = productPriceRepository;
         }
 
-        public async Task<AppResponse<CartDTO>> AddToCart(CartDTO request)
+        public async Task<AppResponse<CartResponse>> AddToCart(CartRequest request)
         {
-            var result = new AppResponse<CartDTO>();
+            var result = new AppResponse<CartResponse>();
             try
             {
                 var user = await _userManager.FindByEmailAsync(_httpContextAccessor.HttpContext?.User.Identity?.Name!);
-                if (user == null)
-                    throw new Exception("User not found.");
-
                 var product = _productRepository.FindByAsync(p => p.Id == request.ProductId).FirstOrDefault();
-                if (product == null)
-                    throw new Exception("Product not found.");
+                if (product == null || product.IsDeleted == true)
+                    return result.BuildError("Product not found or deleted");
 
                 var existingCart = _cartRepository
-                    .FindByAsync(c => c.ProductId == request.ProductId && c.Type == request.Type && c.TenantId == user.TenantId && c.IsDeleted == false)
-                    .FirstOrDefault();
-
+                    .FindByAsync(c => c.ProductId == request.ProductId && c.Type == request.Type
+                    && c.TenantId == user!.TenantId && c.IsDeleted == false).FirstOrDefault();
                 if (existingCart != null)
                 {
                     existingCart.Quantity += request.Quantity;
-                    existingCart.TotalAmount = CalculateTotalAmount(request.ProductId!.Value, request.ProductPriceId!.Value, request.Type, existingCart.Quantity);
+                    existingCart.TotalAmount = CalculateTotalAmount(request.ProductId!.Value, request.ProductPriceId!.Value,
+                        request.Type, existingCart.Quantity);
                     _cartRepository.Delete(existingCart);
-
                     var updatedCart = new Cart
                     {
                         Id = Guid.NewGuid(),
@@ -63,14 +58,19 @@ namespace EasyMall.Services.Implements
                         Type = existingCart.Type,
                         Quantity = existingCart.Quantity,
                         TotalAmount = existingCart.TotalAmount,
-                        TenantId = user.TenantId,
+                        TenantId = user!.TenantId,
                         ProductPriceId = request.ProductPriceId,
                         CreatedBy = user.Email,
                         Modifiedby = user.Email,
                         ModifiedOn = DateTime.UtcNow
                     };
                     _cartRepository.Add(updatedCart);
-                    result.BuildResult(_mapper.Map<CartDTO>(updatedCart));
+
+                    var response = _mapper.Map<CartResponse>(updatedCart);
+                    response.ProductName = product.Name;
+                    var productPrice = _productPriceRepository.FindByAsync(p => p.Id == request.ProductPriceId).FirstOrDefault();
+                    response.ProductPriceName = productPrice?.Type!;
+                    result.BuildResult(response, "Cart updated successfully");
                 }
                 else
                 {
@@ -80,14 +80,20 @@ namespace EasyMall.Services.Implements
                         ProductId = request.ProductId,
                         Type = request.Type,
                         Quantity = request.Quantity,
-                        TotalAmount = CalculateTotalAmount(request.ProductId!.Value, request.ProductPriceId!.Value, request.Type, request.Quantity),
-                        TenantId = user.TenantId,
+                        TotalAmount = CalculateTotalAmount(request.ProductId!.Value, request.ProductPriceId!.Value,
+                            request.Type, request.Quantity),
+                        TenantId = user!.TenantId,
                         ProductPriceId = request.ProductPriceId,
                         CreatedBy = user?.Email,
                         CreatedOn = DateTime.UtcNow
                     };
                     _cartRepository.Add(newCart);
-                    result.BuildResult(_mapper.Map<CartDTO>(newCart), "Added product to cart successfully");
+
+                    var response = _mapper.Map<CartResponse>(newCart);
+                    response.ProductName = product.Name;
+                    var productPrice = _productPriceRepository.FindByAsync(p => p.Id == request.ProductPriceId).FirstOrDefault();
+                    response.ProductPriceName = productPrice?.Type!;
+                    result.BuildResult(response, "Added product to cart successfully");
                 }
             }
             catch (Exception ex)
@@ -100,13 +106,12 @@ namespace EasyMall.Services.Implements
         private double CalculateTotalAmount(Guid productId, Guid productPriceId, string type, int quantity)
         {
             var product = _productRepository.FindByAsync(p => p.Id == productId).FirstOrDefault();
-            if (product == null)
-                throw new Exception("Product not found.");
+            if (product == null || product.IsDeleted == true)
+                throw new Exception("Product not found or deleted");
 
             var productPrice = _productPriceRepository.FindByAsync(p => p.Id == productPriceId && p.Type == type).FirstOrDefault();
-            if (productPrice == null)
-                throw new Exception("Product price not found for the given type.");
-
+            if (productPrice == null || productPrice.IsDeleted == true)
+                throw new Exception("Product price not found or deleted for the given type");
             return (product.Price + productPrice.Price) * quantity;
         }
 
@@ -117,21 +122,21 @@ namespace EasyMall.Services.Implements
             {
                 var user = _httpContextAccessor.HttpContext?.User.Identity?.Name!;
                 var cart = _cartRepository.FindByAsync(c => c.ProductId == productId && c.IsDeleted == false).FirstOrDefault();
-                if (cart == null)
-                    throw new Exception("Product not found in the cart.");
+                if (cart == null || cart.IsDeleted == true)
+                    return result.BuildError("Product not found or deleted in the cart");
                 if (cart.Quantity > 1)
                 {
                     cart.Quantity -= 1;
                     cart.TotalAmount = CalculateTotalAmount(cart.ProductId!.Value, cart.ProductPriceId!.Value, cart.Type, cart.Quantity);
                     _cartRepository.Edit(cart);
-                    result.BuildResult("Product quantity decreased successfully.");
+                    result.BuildResult("Product quantity decreased successfully");
                 }
                 else
                 {
                     cart.Quantity = 0;
                     cart.IsDeleted = true;
                     _cartRepository.Edit(cart);
-                    result.BuildResult("Product removed from cart successfully.");
+                    result.BuildResult("Product removed from cart successfully");
                 }
             }
             catch (Exception ex)
@@ -148,84 +153,17 @@ namespace EasyMall.Services.Implements
             {
                 var user = _httpContextAccessor.HttpContext?.User.Identity?.Name!;
                 var cart = _cartRepository.FindByAsync(c => c.ProductId == productId && c.IsDeleted == false).FirstOrDefault();
-                if (cart == null)
-                    throw new Exception("Product not found in the cart.");
+                if (cart == null || cart.IsDeleted == true)
+                    return result.BuildError("Product not found or deleted in the cart");
                 cart.IsDeleted = true;
                 _cartRepository.Edit(cart);
-                result.BuildResult("Product removed from cart successfully.");
+                result.BuildResult("Product removed from cart successfully");
             }
             catch (Exception ex)
             {
                 result.BuildError(ex.Message);
             }
             return result;
-        }
-
-        public AppResponse<SearchResponse<CartDTO>> Search(SearchRequest request)
-        {
-            var result = new AppResponse<SearchResponse<CartDTO>>();
-            try
-            {
-                var query = BuildFilterExpression(request.Filters!);
-                var numOfRecords = _cartRepository.CountRecordsByPredicate(query);
-                var carts = _cartRepository.FindByPredicate(query);
-                if (request.SortBy != null)
-                    carts = _cartRepository.addSort(carts, request.SortBy);
-                else
-                    carts = carts.OrderBy(x => x.Product!.Name);
-
-                int pageIndex = request.PageIndex ?? 1;
-                int pageSize = request.PageSize ?? 1;
-                int startIndex = (pageIndex - 1) * pageSize;
-                var categoryList = carts.Skip(startIndex).Take(pageSize);
-                var dtoList = _mapper.Map<List<CartDTO>>(categoryList);
-                var searchResponse = new SearchResponse<CartDTO>
-                {
-                    TotalRows = numOfRecords,
-                    TotalPages = CalculateNumOfPages(numOfRecords, pageSize),
-                    CurrentPage = pageIndex,
-                    Data = dtoList,
-                };
-                result.Data = searchResponse;
-                result.IsSuccess = true;
-            }
-            catch (Exception ex)
-            {
-                result.BuildError(ex.Message + " " + ex.StackTrace);
-            }
-            return result;
-        }
-
-        private ExpressionStarter<Cart> BuildFilterExpression(List<Filter> filters)
-        {
-            try
-            {
-                var predicate = PredicateBuilder.New<Cart>(true);
-                if (filters != null)
-                {
-                    foreach (var filter in filters)
-                    {
-                        switch (filter.FieldName)
-                        {
-                            case "Name":
-                                predicate = predicate.And(x => x.Product!.Name.Contains(filter.Value));
-                                break;
-                            case "Price":
-                                predicate = predicate.And(x => x.Product!.Price == double.Parse(filter.Value));
-                                break;
-                            default:
-                                break;
-                        }
-                    }
-                }
-
-                predicate = predicate.And(x => x.IsDeleted == false);
-                return predicate;
-            }
-            catch (Exception ex)
-            {
-                throw new Exception(ex.Message + " " + ex.StackTrace);
-            }
         }
     }
 }
